worked on Simulation functionality

+++ IncohAutoCorrelate/Crystal.cpp
@@ -89,7 +89,7 @@ std::vector<Crystal::Emitter> Crystal::GetEmitters(EmittingCrystSettings Setting
 		EM.Position[1] = At_temp[i][1];
 		EM.Position[2] = At_temp[i][2];
 		if (Settings.Incoherent)
-			EM.Phase = (float) (Drand() * 2 * M_PIl);
+			EM.Phase = (Drand() * 2 * M_PIl);
 	}
 
 	return Ret;
